server/requirements.txt
bidict==0.22.1
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0
eventlet==0.33.3
Faker==18.3.1
Flask==2.2.2
flask-mongoengine==1.0.0
Flask-SocketIO==5.3.3
Flask-WTF==1.1.1
gevent==22.10.2
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
Flask-Cors==4.0.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.8.0
six==1.16.0
Werkzeug==2.2.3
WTForms==3.0.1
zope.event==4.6
zope.interface==6.0
